feat(TraceVisualizer): support trace record visualization (#96)
In this PR, we introduce a new binary `warpo_trace_visualizer` to
convert tracePoint data created by wasm-compiler to perfetto format
- introduce 3rd protozero to generate protobuf2 formatted perfetto trace
- disable `hicpp-no-assembler` clang-tidy rule for use inline assembly

## why protozero

It is lightweight protobuf2 lib to encode and decode.
The C lib is too hard to integration without bugs.
The google original lib is to heavy and has lots deps. And we don't want
a code generate process inside build.

The disadvantage of protozero is we must define the tag manually from
original proto file. Since we don't use too much field. It should be
acceptable.